+++ esp_cam_sensor/sensors/ov3640/include/ov3640_types.h
@@ -1,0 +1,25 @@
+/*
+ * SPDX-FileCopyrightText: 2026 Espressif Systems (Shanghai) CO LTD
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+#pragma once
+
+#include <stdint.h>
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/*
+ * OV3640 camera sensor register type definition.
+ */
+typedef struct {
+    uint16_t reg;
+    uint8_t val;
+} ov3640_reginfo_t;
+
+#ifdef __cplusplus
+}
+#endif